requirements.txt
pydub
flask
openai
python-dotenv
scipy
soundfile
py-espeak-ng